rename useUserLastSeenLimit -> userLastSeenThreshold
